+++ backend/app/models/facture.py
@@ -1,4 +1,5 @@
 from datetime import UTC, datetime
+from decimal import Decimal
 
 from sqlalchemy import DateTime, ForeignKey, Index, Integer, Numeric, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
@@ -19,9 +20,9 @@ class Facture(Base):
     devis_id: Mapped[int] = mapped_column(ForeignKey("devis.id"), nullable=False, index=True)
     numero: Mapped[str] = mapped_column(String(50), nullable=False)
     date_emission: Mapped[datetime] = mapped_column(DateTime, default=lambda: datetime.now(UTC), nullable=False)
-    montant_ht: Mapped[float] = mapped_column(Numeric(10, 2), nullable=False, default=0)
-    tva: Mapped[float] = mapped_column(Numeric(10, 2), nullable=False, default=0)
-    montant_ttc: Mapped[float] = mapped_column(Numeric(10, 2), nullable=False, default=0)
+    montant_ht: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False, default=0)
+    tva: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False, default=0)
+    montant_ttc: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False, default=0)
     status: Mapped[str] = mapped_column(String(30), nullable=False, default="emise", index=True)
     created_at: Mapped[datetime] = mapped_column(DateTime, default=lambda: datetime.now(UTC), nullable=False)
 
@@ -35,7 +36,7 @@ class FactureLigne(Base):
     facture_id: Mapped[int] = mapped_column(ForeignKey("factures.id", ondelete="CASCADE"), nullable=False, index=True)
     designation: Mapped[str] = mapped_column(String(255), nullable=False)
     quantite: Mapped[int] = mapped_column(Integer, nullable=False, default=1)
-    prix_unitaire_ht: Mapped[float] = mapped_column(Numeric(10, 2), nullable=False, default=0)
-    taux_tva: Mapped[float] = mapped_column(Numeric(5, 2), nullable=False, default=20.0)
-    montant_ht: Mapped[float] = mapped_column(Numeric(10, 2), nullable=False, default=0)
-    montant_ttc: Mapped[float] = mapped_column(Numeric(10, 2), nullable=False, default=0)
+    prix_unitaire_ht: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False, default=0)
+    taux_tva: Mapped[Decimal] = mapped_column(Numeric(5, 2), nullable=False, default=20.0)
+    montant_ht: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False, default=0)
+    montant_ttc: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False, default=0)
